+++ src/Delete.java
@@ -2,5 +2,5 @@ import java.util.List;
 import java.util.Map;
 
 public interface Delete {
-    public String remove(String name, int quantity, Map<String, List<Product>> productList);
+    public String remove(String name, double quantity, Map<String, List<Product>> productList);
 }
