+++ src/ManageHistory.java
@@ -1,4 +1,4 @@
 public interface ManageHistory {
-    void addNewChange(String date,String name,int quantity);
+    void addNewChange(String date,String name,double quantity);
     void getHistoryInfo(String fromDate,String toDate);
 }
